+++ crates/teepot-tee-quote-verification-rs/src/empty.rs
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright (c) 2025 Matter Labs
+
+pub const NOTHING_TO_SEE_HERE: u8 = 0;
