+++ crates/teepot/src/quote/mod.rs
@@ -6,6 +6,7 @@
 
 //! Get a quote from a TEE
 
+#[cfg(feature = "quote_op")]
 pub mod attestation;
 pub mod error;
 pub mod tcblevel;
@@ -15,6 +16,7 @@ pub mod tcblevel;
     not(all(target_os = "linux", target_arch = "x86_64")),
     path = "phala.rs"
 )]
+#[cfg(feature = "quote_op")]
 mod os;
 mod utils;
 
@@ -644,6 +646,7 @@ impl FromStr for TEEType {
 }
 
 /// Get the attestation quote from a TEE
+#[cfg(feature = "quote_op")]
 pub fn get_quote(report_data: &[u8]) -> Result<(TEEType, Box<[u8]>), QuoteError> {
     os::get_quote(report_data)
 }
@@ -690,11 +693,13 @@ pub struct Collateral {
 }
 
 /// Get the collateral data from an SGX or TDX quote
+#[cfg(feature = "quote_op")]
 pub fn get_collateral(quote: &[u8]) -> Result<Collateral, QuoteError> {
     os::get_collateral(quote)
 }
 
 /// Verifies a quote with optional collateral material
+#[cfg(feature = "quote_op")]
 pub fn verify_quote_with_collateral(
     quote: &[u8],
     collateral: Option<&Collateral>,
